moq-mux: extract the H.264 byte->frame Split; add the FrameDecode path
First slice of the splitter / unified-decode work.

- New `codec::h264::Split`: a standalone parser that turns H.264 bytes into
  `container::Frame`s plus a resolved `VideoConfig` (the NAL/AU assembly, avcC
  parsing, SPS/PPS cache, and Annex-B wall-clock all live here, independently
  testable). avc1 still errors on reconfiguration; avc3 still updates in place.
- `codec::h264::Import` now drives `Split` internally for its byte APIs
  (`decode_frame` / `decode_stream` unchanged, so the Framed/Stream/TS
  dispatchers and external callers are untouched and don't regress) and pulls
  the resolved config into its local catalog.
- New `publish::FrameDecode` trait (`decode(impl IntoIterator<Item = Frame>)`),
  the uniform decode entry point a caller drives with frames from its own
  `Split`. `Published::decode` wraps it and syncs, so the catalog re-mirror
  can't be forgotten — the footgun-free path. h264::Import implements it.

The dispatchers still use the byte API + manual `sync()`; migrating them to
`Split` + `Published::decode` (and resolving the avc1-vs-avc3 config flow on
that path) is the next slice, then rolling the same split to the other codecs.
