+++ rs/moq-mux/src/codec/h264/mod.rs
@@ -10,9 +10,11 @@
 
 mod export;
 mod import;
+mod split;
 
 pub use export::*;
 pub use import::*;
+pub use split::*;
 
 use bytes::{Buf, BufMut, Bytes, BytesMut};
